refactor(config): reorganize configuration modules and enhance structure
* Moved `run_migrations` and `migrate_legacy_config` functions to `database.py`.
* Created `base.py` for persistent configuration management.
* Updated `__init__.py` to reflect new module structure and maintain functionality.
