+++ backend/open_webui/config/database.py
@@ -1,0 +1,85 @@
+from datetime import datetime
+import json
+import logging
+import os
+
+from open_webui.env import (
+    DATA_DIR,
+    OPEN_WEBUI_DIR
+)
+from sqlalchemy import (
+    JSON,
+    Column,
+    DateTime,
+    Integer,
+    func
+)
+from open_webui.internal.db import Base, get_db
+
+log = logging.getLogger(__name__)
+
+DEFAULT_CONFIG = {
+    "version": 0,
+    "ui": {},
+}
+
+def run_migrations():
+    log.info("Running migrations")
+    try:
+        from alembic import command
+        from alembic.config import Config
+
+        alembic_cfg = Config(OPEN_WEBUI_DIR / "alembic.ini")
+
+        # Set the script location dynamically
+        migrations_path = OPEN_WEBUI_DIR / "migrations"
+        alembic_cfg.set_main_option("script_location", str(migrations_path))
+
+        command.upgrade(alembic_cfg, "head")
+    except Exception as e:
+        log.exception(f"Error running migrations: {e}")
+
+
+class Config(Base):
+    __tablename__ = "config"
+
+    id = Column(Integer, primary_key=True)
+    data = Column(JSON, nullable=False)
+    version = Column(Integer, nullable=False, default=0)
+    created_at = Column(DateTime, nullable=False, server_default=func.now())
+    updated_at = Column(DateTime, nullable=True, onupdate=func.now())
+
+def load_json_config():
+    with open(f"{DATA_DIR}/config.json", "r") as file:
+        return json.load(file)
+
+
+def save_to_db(data):
+    with get_db() as db:
+        existing_config = db.query(Config).first()
+        if not existing_config:
+            new_config = Config(data=data, version=0)
+            db.add(new_config)
+        else:
+            existing_config.data = data
+            existing_config.updated_at = datetime.now()
+            db.add(existing_config)
+        db.commit()
+
+def reset_config():
+    with get_db() as db:
+        db.query(Config).delete()
+        db.commit()
+
+
+def migrate_legacy_config():
+  if os.path.exists(f"{DATA_DIR}/config.json"):
+      data = load_json_config()
+      save_to_db(data)
+      os.rename(f"{DATA_DIR}/config.json", f"{DATA_DIR}/old_config.json")
+
+
+def get_config():
+    with get_db() as db:
+        config_entry = db.query(Config).order_by(Config.id.desc()).first()
+        return config_entry.data if config_entry else DEFAULT_CONFIG
